requirements.txt
joblib==1.0.1
numpy==1.20.1
pandas==1.2.3
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
scikit-learn==0.24.1
scipy==1.6.2
six==1.15.0
threadpoolctl==2.1.0
